requirements.txt
asgiref==3.4.1
autopep8==1.5.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
